requirements.txt
anthropic>=0.25.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
reportlab>=4.0.0
python-dotenv>=1.0.0
tqdm>=4.66.0
